New sfpu_domains.py file added that can exclude undefined intervals for certain sfpu operations. for_op() moved to that file.
