+++ tt_metal/tt-llk/tests/python_tests/helpers/stimuli_generator_v2.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 # SPDX-License-Identifier: Apache-2.0
 
-import copy
 import math
 import warnings
 from dataclasses import dataclass
@@ -16,7 +15,7 @@ from .format_config import (
     MXFP8_E5M2_MAX_NORMAL,
     DataFormat,
 )
-from .llk_params import MathOperation, format_dict
+from .llk_params import format_dict
 from .tile_constants import (
     DEFAULT_TILE_C_DIM,
     DEFAULT_TILE_R_DIM,
@@ -393,322 +392,6 @@ class StimuliSpec:
             **kwargs,
         )
 
-    @classmethod
-    def for_op(
-        cls,
-        op: MathOperation,
-        data_format: DataFormat = DataFormat.Float16_b,
-    ) -> "OperandSpecs":
-        """Return OperandSpecs with safe input domains for *op* and *data_format*.
-
-        Args:
-            op: Target math operation.
-            data_format: Input data format; controls the numeric range and
-            precision used to choose safe per-op input domains (e.g. tighter
-            ranges for narrower MX/BFP formats).
-
-        Returns:
-            OperandSpecs with per-operand domain specs.
-
-        Raises:
-            KeyError: If *op* is not in the registry.
-        """
-        # Deep-copy so callers can safely mutate specs without altering the registry.
-        entry = _OP_DOMAIN_REGISTRY.get(op)
-        if entry is None:
-            registered = sorted(o.name for o in _OP_DOMAIN_REGISTRY)
-            raise KeyError(
-                f"MathOperation.{op.name} has no entry in the stimuli domain "
-                f"registry.  Add an OperandSpecs entry to _OP_DOMAIN_REGISTRY.\n"
-                f"Currently registered ({len(registered)}): {registered}"
-            )
-        if callable(entry):
-            return copy.deepcopy(entry(data_format))
-        return copy.deepcopy(entry)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# OperandSpecs
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class OperandSpecs:
-    """Per-operand input domain specs returned by StimuliSpec.for_op method.
-
-    For binary ops where operands need different domains (e.g. divisor avoids
-    zero), spec_A and spec_B differ; unary ops need only spec_A.
-    spec_B defaults to a copy of spec_A when "None".
-    """
-
-    spec_A: StimuliSpec
-    spec_B: Optional[StimuliSpec] = None
-
-    def __post_init__(self) -> None:
-        if self.spec_B is None:
-            self.spec_B = copy.deepcopy(self.spec_A)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Format-specific domain builders (referenced by _OP_DOMAIN_REGISTRY)
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def _exp_spec(fmt: DataFormat) -> OperandSpecs:
-    """Safe input range for exp(x) per format to avoid overflow."""
-    if fmt == DataFormat.MxFp8P:
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    elif fmt in (DataFormat.Float16, DataFormat.MxFp8R):
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    else:
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-80.0, high=80.0)
-    return OperandSpecs(spec_A=spec)
-
-
-def _exp2_spec(fmt: DataFormat) -> OperandSpecs:
-    """Safe input range for exp2(x) = 2^x per format to avoid overflow."""
-    if fmt == DataFormat.MxFp8P:
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-7.0, high=7.0)
-    elif fmt in (DataFormat.Float16, DataFormat.MxFp8R):
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-14.0, high=14.0)
-    else:
-        spec = StimuliSpec(
-            distribution=DistributionKind.UNIFORM, low=-100.0, high=100.0
-        )
-    return OperandSpecs(spec_A=spec)
-
-
-def _square_spec(fmt: DataFormat) -> OperandSpecs:
-    """Safe input range for square(x) = x^2 per format to avoid overflow."""
-    if fmt == DataFormat.MxFp8P:
-        spec = StimuliSpec(distribution=DistributionKind.UNIFORM, low=-20.0, high=20.0)
-    elif fmt in (DataFormat.Float16, DataFormat.MxFp8R):
-        spec = StimuliSpec(
-            distribution=DistributionKind.UNIFORM, low=-200.0, high=200.0
-        )
-    else:
-        spec = StimuliSpec(
-            distribution=DistributionKind.UNIFORM, low=-1000.0, high=1000.0
-        )
-    return OperandSpecs(spec_A=spec)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# SFPU / FPU operation domain registry
-# ─────────────────────────────────────────────────────────────────────────────
-#
-# Maps every MathOperation to either:
-#   OperandSpecs          — format-independent safe input domains
-#   callable              — (DataFormat) -> OperandSpecs for format-sensitive ops
-#
-# For unary operations spec_B is omitted (defaults to a copy of spec_A).
-# For binary operations where operands require different domains the entry
-# uses explicit spec_A and spec_B.
-#
-# Rationale for each entry is documented inline.
-
-_OP_DOMAIN_REGISTRY: Dict[
-    MathOperation,
-    Union[OperandSpecs, Callable[[DataFormat], OperandSpecs]],
-] = {
-    # ── SFPU unary ────────────────────────────────────────────────────────────
-    # abs: all reals; include negative branch
-    MathOperation.Abs: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    # acosh: domain x >= 1
-    MathOperation.Acosh: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=1.0, high=10.0)
-    ),
-    # asinh: all reals
-    MathOperation.Asinh: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    # atanh: domain |x| < 1; stay away from ±1 to avoid ±inf
-    MathOperation.Atanh: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-0.95, high=0.95)
-    ),
-    # celu: exercises both the exponential branch (x < 0) and linear (x >= 0)
-    MathOperation.Celu: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    # cos: cover the full unit circle
-    MathOperation.Cos: OperandSpecs(
-        spec_A=StimuliSpec(
-            distribution=DistributionKind.UNIFORM, low=-math.pi, high=math.pi
-        )
-    ),
-    # elu: exercises the exponential branch (x < 0)
-    MathOperation.Elu: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    # exp: format-specific overflow threshold
-    MathOperation.Exp: _exp_spec,
-    # exp2: format-specific overflow threshold
-    MathOperation.Exp2: _exp2_spec,
-    # fill: the hardware ignores the input value; any range is fine
-    MathOperation.Fill: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=1.0)
-    ),
-    # gelu: Gaussian activation — gaussian distribution naturally exercises tails
-    MathOperation.Gelu: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.GAUSSIAN, mean=0.0, std=3.0)
-    ),
-    # hardsigmoid: linear region between -3 and 3, clipped outside
-    MathOperation.Hardsigmoid: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-4.0, high=4.0)
-    ),
-    # log: domain x > 0; log-uniform spans several decades
-    MathOperation.Log: OperandSpecs(
-        spec_A=StimuliSpec(
-            distribution=DistributionKind.LOG_UNIFORM, low=1e-4, high=1e3
-        )
-    ),
-    # log1p: domain x > -1; log1p(x) = log(1 + x)
-    MathOperation.Log1p: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-0.99, high=10.0)
-    ),
-    # neg: all reals
-    MathOperation.Neg: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    # reciprocal: domain x != 0; log-uniform avoids zero and covers decades.
-    # Only strictly positive values are generated here so that negative-
-    # reciprocal paths are intentionally not exercised by default.
-    # To test negative inputs, supply an explicit spec such as:
-    #   StimuliSpec.uniform(low=-100.0, high=-0.1)
-    MathOperation.Reciprocal: OperandSpecs(
-        spec_A=StimuliSpec(
-            distribution=DistributionKind.LOG_UNIFORM, low=0.1, high=100.0
-        )
-    ),
-    # relu / relu_max / relu_min / threshold: include negatives (zero branch)
-    MathOperation.Relu: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    MathOperation.ReluMax: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    MathOperation.ReluMin: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    MathOperation.Threshold: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    # rsqrt: domain x > 0; log-uniform covers a wide positive range
-    MathOperation.Rsqrt: OperandSpecs(
-        spec_A=StimuliSpec(
-            distribution=DistributionKind.LOG_UNIFORM, low=1e-4, high=100.0
-        )
-    ),
-    # sigmoid: cover both saturation regions
-    MathOperation.Sigmoid: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-8.0, high=8.0)
-    ),
-    # silu: silu(x) = x * sigmoid(x); cover saturation + linear regions
-    MathOperation.Silu: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    # sin: cover the full unit circle
-    MathOperation.Sin: OperandSpecs(
-        spec_A=StimuliSpec(
-            distribution=DistributionKind.UNIFORM, low=-math.pi, high=math.pi
-        )
-    ),
-    # sqrt: domain x >= 0
-    MathOperation.Sqrt: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=100.0)
-    ),
-    # square: format-specific overflow threshold
-    MathOperation.Square: _square_spec,
-    # tanh: cover saturation regions (saturates near ±1 for |x| > ~3)
-    MathOperation.Tanh: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0)
-    ),
-    # topk family: operation sorts/merges; any values are valid
-    MathOperation.TopKLocalSort: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    MathOperation.TopKMerge: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    MathOperation.TopKRebuild: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-10.0, high=10.0)
-    ),
-    # ── FPU binary ────────────────────────────────────────────────────────────
-    MathOperation.Elwadd: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.Elwmul: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.Elwsub: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    # ── SFPU binary ───────────────────────────────────────────────────────────
-    MathOperation.SfpuElwadd: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.SfpuElwmul: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.SfpuElwsub: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    # div: srcA is the dividend (any value); srcB is the divisor.
-    # The divisor uses log-uniform over strictly positive values so that
-    # divide-by-zero and near-zero instability are avoided.
-    # This means negative-divisor paths are intentionally not exercised here.
-    # To test negative-divisor behaviour, supply an explicit spec_B such as:
-    #   StimuliSpec.uniform(low=-10.0, high=-0.1)
-    MathOperation.SfpuElwdiv: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-2.0, high=2.0),
-        spec_B=StimuliSpec(
-            distribution=DistributionKind.LOG_UNIFORM, low=0.1, high=10.0
-        ),
-    ),
-    MathOperation.SfpuElwrsub: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    # pow: srcA is the base (must be non-negative for non-integer exponents);
-    # srcB is the exponent (non-negative to keep output finite)
-    MathOperation.SfpuElwpow: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=5.0),
-        spec_B=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=5.0),
-    ),
-    # xlogy: computes x * log(y) element-wise
-    # srcA (x): x >= 0 so xlogy(0, y) = 0 is well-defined
-    # srcB (y): y > 0 so log(y) is finite; log-uniform spans several decades
-    MathOperation.SfpuXlogy: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=10.0),
-        spec_B=StimuliSpec(
-            distribution=DistributionKind.LOG_UNIFORM, low=1e-4, high=10.0
-        ),
-    ),
-    MathOperation.SfpuAddTopRow: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    # shift ops: operate on integer bit patterns; both operands in [0, 255]
-    MathOperation.SfpuElwLeftShift: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=255.0)
-    ),
-    MathOperation.SfpuElwLogicalRightShift: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=255.0)
-    ),
-    MathOperation.SfpuElwRightShift: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=0.0, high=255.0)
-    ),
-    # ── Reduce ────────────────────────────────────────────────────────────────
-    MathOperation.ReduceColumn: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.ReduceRow: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-    MathOperation.ReduceScalar: OperandSpecs(
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-1.0, high=1.0)
-    ),
-}
 
 # ─────────────────────────────────────────────────────────────────────────────
 # Tile / face count utilities
